% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{accumulation_level}
\alias{accumulation_level}
\alias{plot.accumulation_level}
\title{Correlation between a chronology and climate data at different
  accumulation levels}
\usage{
accumulation_level(crn, climate.data, accumulation, accumfun = sum,
  months = c(-2:10), method = "spearman")

\method{plot}{accumulation_level}(accumulation_level_object,
  main = "accumulation level")
}
\arguments{
\item{crn}{A chronology - the first column is used by the function, make
sure to give the right input (in case multiple chronologies are in one
object, e.g. as for AR-chronologies).}

\item{climate.data}{climate data in "wide format" with years in the first
column and monthly climate measurement in the consecutive 12 columns.}

\item{accumulation}{an integer specifying up to which window size
accumulations get calculated.}

\item{accumfun}{function to use for accumulation of the climate measure
within the window.}

\item{months}{an integer sequence selecting the end months of the windows to
use. 1:5 selects jan to may of the "current" years as end months. to select
months of the "previous" years use negative integers - c(-2:10) selects last
years October to current years October as end months. Selection needs to be
within c(-11:12).}

\item{method}{method to calculate correlations - "pearson", "kendall"
or "spearman".}

\item{accumulation_level_object}{an object obtained by accumulation_level.}

\item{main}{a plot title}
}
\value{
a matrix with the same number of rows as selected aggregation levels
  (argument "aggregation") and columns as the selection of end-months.
}
\description{
The climate data is accumulated (eg. summed or averaged) over
  varying window sizes (from 1 to window size set with "accumulation")
  and correlations with a given chronology are calculated ("spearman" by
  default, but can be set).
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot the output of accumulation_level()
as contour plot.
}}

\examples{
# not available in development version
}
