% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{afrp}
\alias{afrp}
\title{adjusted false ring proportion}
\usage{
afrp(x)
}
\arguments{
\item{x}{A data frame with numeric columns representing individual series
and years as rownames where years with IADF are marked with 1,
those without with 0, years not covered by the series are set to NA.}
}
\value{
double, years as names
}
\description{
This function calulates the adjusted false fing proportion, as
suggested by Osborn et. al. (1997), of a set of binary false ring assignments
}
\references{
Osborn TJ, Briffa KR and Jones PD (1997) Adjusting variance for
  sample-size in tree-ring chronologies and other regional-mean timeseries.
  Dendrochronologia 15, 89-99.
}
\seealso{
\code{\link{frp}}
}

