% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono-quality.R
\name{age_class_plot}
\alias{age_class_plot}
\title{Plot median growth rate or other parameters per age class}
\usage{
age_class_plot(dat, cutoff = NULL, classborders = c(0, 60, 80, 100,
  150, Inf), main = "", unit = "[mm]")
}
\arguments{
\item{dat}{data of the investigated tree ring parameter (ring width,
earlywood width, leatewood with, ring density, etc.).}

\item{cutoff}{cut the data at given year - if NULL the whole range is used.}

\item{classborders}{cambial ages to set the borders of the age classes.}

\item{main}{plot title.}

\item{unit}{y-axis label.}
}
\description{
Plot median growth rate or other parameters per age class to
  evaluate trends in tree ring data and data composition.
}
\examples{
#not available in development version.
}
