% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{age_rwl}
\alias{age_rwl}
\title{age_rwl}
\usage{
age_rwl(rwl)
}
\arguments{
\item{rwl}{a rwl/data.frame object}
}
\value{
a rwl/data.frame object with cambial age of the years of the series.
}
\description{
A function returning an rwl/data.frame object with the same
  dimensions as rwl with showing the cambial age instead of tree ring width
  in the corresponding year.
}
\examples{
library("dplR")
data("gp.rwl")
age_rwl(gp.rwl)
}

