% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono-quality.R
\name{agegrowthplot}
\alias{agegrowthplot}
\title{agegrowthplot}
\usage{
agegrowthplot(rwl, po, subset = NULL, subset.color = "violet", main = "")
}
\arguments{
\item{rwl}{an rwl object}

\item{po}{a data frame with series names in the first and pith offset values
(number of rings) in the second column}

\item{subset}{provide a character string of names in rwl which get
a different color in the last plot}

\item{subset.color}{color of the subset}

\item{main}{optional plot title}
}
\description{
plots mean annual growth rate in a common interval (shown in
  first plot) over tree age - a subset (eg. historic vs. recent trees,
  a different site ect.) can be specified to get a different color in the plot.
}
\examples{
library("dplR")
data("gp.rwl")
data("gp.po")
sub <- names(gp.rwl)[substr(names(gp.rwl), 3, 3) == "B"]
agegrowthplot(gp.rwl, gp.po, subset = sub, main = 'my age-growth-plot')
}

