% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{avg_series_profiles}
\alias{avg_series_profiles}
\title{average rescaled density profiles}
\usage{
avg_series_profiles(resc)
}
\arguments{
\item{resc}{a nested list as obtained from \code{\link{rescale_series_profiles}}}
}
\value{
a list with the members $profile.raw, $limites and $years
}
\examples{
#not provided in the development version
}

