% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{catchment}
\alias{catchment}
\title{find upstream catchments}
\usage{
catchment(catchments_digHAO, lonlat, proj4string = sp::CRS("+init=epsg:4326"),
  make.plot = TRUE, rivers_digHAO, ...)
}
\arguments{
\item{catchments_digHAO}{a SpatialPolygonsDataFrame imported with
\code{\link[maptools]{readShapePoly}}, data covered by the digital version
of the hydrological atlas of austria}

\item{lonlat}{a numeric vector of length 2 with the longitue and latitude of a mappoint}

\item{proj4string}{the proj4string of the provided lonlat}

\item{make.plot}{logic}

\item{rivers_digHAO}{a SpatialLinesDataFrame imported with
\code{\link[maptools]{readShapeLines}}, data covered by the digital version
of the hydrological atlas of austria}

\item{...}{additional plotting arguments such as plot title etc.}
}
\value{
a list with index values as first and HYDROIDs as second member
}
\description{
find all upstream catchments given coordinates of a point within
  a catchment covered by digHAO (digital version of the hydrological atlas of
  austria)
}
\examples{
\dontrun{
 path_digHAO <- 'ENTER PATH TO YOUR digHAO DIRECTORY HERE'

 rivers <- maptools::readShapeLines(paste0(path_digHAO,
 '/Shape/Gemeinsam/gew1mio.shp')) # Lambert-projection Austria
 sp::proj4string(rivers) <- sp::CRS("+init=epsg:31287")

 catchments <- maptools::readShapePoly(paste0(path_digHAO,
 '/Shape/1_3/wasserbilanz.shp')) # Lambert-projection Austria
 sp::proj4string(catchments) <- sp::CRS("+init=epsg:31287")

 lonlat <- c(16, 48.5)
 catchment(catchments, lonlat, proj4string = sp::CRS("+init=epsg:4326"),
 make.plot = TRUE, rivers_digHAO = rivers)
 }
}
\references{
Fürst, J. (2008): Der digitale Hydrologische Atlas Österreichs.
  Der Hydrologische Atlas Österreichs: Anwendung in Wasserwirtschaft und
  Umweltschutz, 1. Juli 2008, Universität für Bodenkultur Wien
}
