% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{catchment_convenient}
\alias{catchment_convenient}
\title{find upstream catchments}
\usage{
catchment_convenient(path_digHAO, lonlat,
  proj4string = sp::CRS("+init=epsg:4326"), make.plot = TRUE, ...)
}
\arguments{
\item{path_digHAO}{path to the digHAO directory as character string}

\item{lonlat}{a numeric vector of length 2 with the longitue and latitude of a mappoint}

\item{proj4string}{the proj4string of the provided lonlat}

\item{make.plot}{logic}

\item{...}{additional plotting arguments such as plot title etc.}
}
\value{
a list with index values as first and HYDROIDs as second member
}
\description{
convenient, but slower, version of \code{\link{catchment}} where
  only the path to the digHAO folder needs to be supplied. If you want to find
  catchments for multiple points in a loop use \code{\link{catchment}} instead
  as catchment_convenient() loads the digHAO data at each function call.
}
\examples{
\dontrun{
 path_digHAO <- 'ENTER PATH TO YOUR digHAO DIRECTORY HERE'

 lonlat <- c(16, 48.5)
 catchment_convenient(path_digHAO, lonlat)
 }
}
\references{
Fürst, J. (2008): Der digitale Hydrologische Atlas Österreichs.
  Der Hydrologische Atlas Österreichs: Anwendung in Wasserwirtschaft und
  Umweltschutz, 1. Juli 2008, Universität für Bodenkultur Wien
}
