% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_fragments.R
\name{combine_fragments}
\alias{combine_fragments}
\title{combine fragments}
\usage{
combine_fragments(rwl, stc = c(3, 4, 1), rules = c(a = "[af-l]$", b =
  "[bm-w]$", c = "[c]$", d = "[dx-z]$"))
}
\arguments{
\item{rwl}{a rwl/data.frame object.}

\item{stc}{numeric, parameter as defined in \code{\link[dplR]{read.ids}}.}

\item{rules}{combining rules as a vector of named regular expressions -
the default combines fragments ending with "f-l" to "a", "m-w" to "b",
"c" to "c" and "x-z" to "d".}
}
\value{
a data.frame.
}
\description{
combine multiple fragments of a series as from measuring parts
  of densitometry samples in WinDENDRO according to set rules - this function
  also averages overlapping measurements.
}
\examples{
rwl <- data.frame('AbcAA01f' = c(1, 1, NA, NA, NA),
                  'AbcAA01g' = c(NA, NA, 1, 1, NA),
                  'AbcAA01h' = c(NA, NA, NA, NA, 1),
                  'AbcAA01m' = c(2, 2, 2, NA, NA),
                  'AbcAA01n' = c(NA, NA, NA, 2, 2),
combine_fragments(rwl)
}

