% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatology.R
\name{dcc_spear}
\alias{dcc_spear}
\alias{plot.dcc_spear}
\title{correlation analysis for tree ring chronologies and climate data}
\usage{
dcc_spear(crn, clim, months = c(-2:10), method = c("range", "mean", "sum"),
  replicates = 1000, conf = 0.95, ci.type = "basic")

\method{plot}{dcc_spear}(x, ...)
}
\arguments{
\item{crn}{A chronology - the first column is used by the function, make
sure to give the right input (in case multiple chronologies are in one
object, e.g. as for AR-chronologies).}

\item{clim}{climate data in "wide format" with years in the first
column and monthly climate measurement in the consecutive 12 columns.}

\item{months}{an integer sequence selecting the end months of the windows to
use. 1:5 selects jan to may of the "current" years as end months. to select
months of the "previous" years use negative integers - c(-2:10) selects last
years October to current years October as end months. Selection needs to be
within c(-11:12).}

\item{method}{a character string, 'range', 'mean' or 'sum', selecting whether
to calculate the correlation with individual months, or with means or sums
over all (by argument month) selected months of the climate data.}

\item{replicates}{integer specifying the number of bootstrap replicates used
to calculate the bootstrapped confidence interval.}

\item{conf}{the confidence level to be used for the
\code{\link[stats]{cor.test}} (represented by column p_value in the output
and the bootstrapped confidence interval computation).}

\item{ci.type}{type of confidence interval to be computed as defined as
"type" in \code{\link[boot]{boot.ci}}.}

\item{x}{an output object from function dcc_spear()}

\item{...}{optional plotting arguments}
}
\value{
a data frame with following names:
  \describe{
   \item{coef}{spearmans rho}
   \item{p_value}{p-values from calling \code{\link[stats]{cor.test}} - that
     is without taking multiplicity into account}
   \item{ci_lower}{lower bound of bootstrapped confidence interval}
   \item{ci_upper}{upper bound of bootstrapped confidence interval}
  }
}
\description{
A function similar to the functionality of \code{\link[treeclim]{dcc}};
  without moving windows, but spearman correlations used.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: plot method for object of class "dcc_spear"
}}

\examples{
#not available in development version
}
