% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{error_bar}
\alias{error_bar}
\title{error bar}
\usage{
error_bar(x, y, ll, ul, width)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate - a point gets drawn there}

\item{ll}{lower limit - position of the lower whisker}

\item{ul}{upper limit - position of the upper whisker}

\item{width}{width of the whisker ends}
}
\description{
add a single error bar to a plot
}
\examples{
#single:
plot(NULL, xlim = c(0, 2), ylim = c(0, 5))
error_bar(1, 2.5, 1, 4.5, 1)

#multiple:
percentile <- do.call('rbind', lapply(list(c(1:4), c(2:5), c(3:6)), quantile))
plot(NULL, xlim = c(0, 4), ylim = c(0, 7))
lapply(seq_len(nrow(percentile)),
       FUN = function(x) error_bar(x,
                                   percentile[x, 3],
                                   percentile[x, 2],
                                   percentile[x, 4],
                                   width = 0.5))
}
