% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{expand_apply.data.frame}
\alias{expand_apply.data.frame}
\title{expand_apply for data frames}
\usage{
expand_apply.data.frame(x, FUN = "median")
}
\arguments{
\item{x}{a data.frame/rwl object}

\item{FUN}{character, name of a function e.g. 'median'.}
}
\examples{
library('dplR')
data('ca533')
expand_apply.data.frame(ca533, 'median')
}

