% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{expand_apply.numeric}
\alias{expand_apply.numeric}
\title{expand_apply}
\usage{
expand_apply.numeric(x, FUN = "median")
}
\arguments{
\item{x}{a numeric vector, NA is allowed and will be omitted.}

\item{FUN}{character, name of a function e.g. 'median'.}
}
\description{
Apply function on expanding window.
}
\examples{
x <- c(NA, NA, 1, 3, 2, 3, NA)
expand_apply.numeric(x, 'median')
}

