% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-filled_contour.R
\name{filled_contour}
\alias{filled_contour}
\title{filled_contour}
\usage{
filled_contour(x = seq(0, 1, length.out = nrow(z)), y = seq(0, 1, length.out
  = ncol(z)), z, xlim = range(x, finite = TRUE), ylim = range(y, finite =
  TRUE), zlim = range(z, finite = TRUE), levels = pretty(zlim, nlevels),
  nlevels = 20, color.palette = cm.colors,
  col = color.palette(length(levels) - 1), plot.title, plot.axes, key.title,
  key.axes, asp = NA, xaxs = "i", yaxs = "i", las = 1, axes = TRUE,
  frame.plot = axes, ...)
}
\arguments{
\item{x, y, z, xlim, ylim, zlim, levels, nlevels, color.palette, col, plot.title, plot.axes, key.title, key.axes, asp, xaxs, yaxs, las, axes, frame.plot, ...}{refer to the documantation of the original function \code{\link[graphics]{filled.contour}}.}
}
\description{
mostly the same as \code{\link[graphics]{filled.contour}},
  \code{border = 0} is set in \code{rect()} as adviced at \url{http://stackoverflow.com/questions/8068366/removing-lines-within-filled-contour-legend}
}

