% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{first_last}
\alias{first_last}
\title{first_last}
\usage{
first_last(x)
}
\arguments{
\item{x}{a data.frame/rwl object}
}
\value{
a data.frame with series names in the first column as character strings,
  and the first as well as the last years of the series in the second resp.
  third column
}
\description{
function to return the first and the last year of tree ring series
 in an rwl object
}
\examples{
library('dplR')
data('gp.rwl')
first_last(gp.rwl)
}

