% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{frp}
\alias{frp}
\title{false ring proportion}
\usage{
frp(x)
}
\arguments{
\item{x}{A data frame with numeric columns representing individual series
and years as rownames where years with IADF are marked binary with 1,
those without with 0, years not covered by the series are set to NA.}
}
\value{
double, years as names
}
\description{
This function calulates the false fing proportion of a set of
  binary false ring assignments
}
\seealso{
\code{\link{afrp}}
}

