% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{iadf_averaged_barplot}
\alias{iadf_averaged_barplot}
\title{iadf_averaged_barplot}
\usage{
iadf_averaged_barplot(iadf, top_space = 3)
}
\arguments{
\item{iadf}{the output of \code{\link{iadf_series_ewlw}}.}

\item{top_space}{numeric, space to add at the top of the graph for nicer
plotting.}
}
\description{
Plot the distribution of the detected IADF intensity by adding
 error bars spanning the IQR to the barplots of the median IADF intensity.
}
\examples{
#not provided in development version.
}

