% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{iadf_multiple_barplot}
\alias{iadf_multiple_barplot}
\title{individual IADF barplots for multiple profiles}
\usage{
iadf_multiple_barplot(resc, iadf_output_list)
}
\arguments{
\item{resc}{a nested list as obtained from \code{\link{rescale_series_profiles}}}

\item{iadf_output_list}{the output of \code{\link{iadf_multiple_ewlw}}}
}
\description{
apply \code{\link{iadf_profile_barplot}} to multiple profiles in
  a list and receive individual boxplots for each.
}
\examples{
#not provided in development version
}

