% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{iadf_multiple_ewlw}
\alias{iadf_multiple_ewlw}
\title{iadf_multiple_ewlw}
\usage{
iadf_multiple_ewlw(resc, make.plot = FALSE, trsh.adjust = 0.003)
}
\arguments{
\item{resc}{a rescaled nested list as obtained from \code{\link{rescale_series_profiles}}}

\item{make.plot}{logic}

\item{trsh.adjust}{an arbitrary threshold to avoid including the potentially
concave beginning of the late wood profile to the iadf measure}
}
\value{
a named list of data frames with "year" in the first column, the
  measured area between a convex approximation and the profile for earlywood
  ('EW') in the second and for latewood ('LW') in the third column.
}
\description{
a wrapper to \code{\link{iadf_series_ewlw}} to work with a
  nested list of multiple series
}
\examples{
#not provided in development version
}

