% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{iadf_profile_barplot}
\alias{iadf_profile_barplot}
\title{iadf_profile_barplot}
\usage{
iadf_profile_barplot(dprofile, iadf_output, main_ew = NULL, main_lw = NULL,
  ...)
}
\arguments{
\item{dprofile}{a list of an individual rescaled series (e.g. a single member
of a list obtained from \code{\link{rescale_series_profiles}}) with the
members $profile.raw, $limites and $years.}

\item{iadf_output}{the output of \code{\link{iadf_series_ewlw}} or a single
member of the output of \code{\link{iadf_multiple_ewlw}}}

\item{main_ew}{a plot title for the earlywood iadf plot}

\item{main_lw}{a plot title for the latewood iadf plot}

\item{...}{other graphical parameters}
}
\description{
plot the density profile together with a barplot of the iadf
  measure
}
\examples{
# not provided in development version
}

