% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{iadf_series_ewlw}
\alias{iadf_series_ewlw}
\title{iadf_series_ewlw}
\usage{
iadf_series_ewlw(dprofile, make.plot = FALSE, trsh.adjust = 0.003)
}
\arguments{
\item{dprofile}{a list of an individual rescaled series (e.g. a single member
of a list obtained from \code{\link{rescale_series_profiles}}) with the
members $profile.raw, $limites and $years.}

\item{make.plot}{logic}

\item{trsh.adjust}{an arbitrary threshold to avoid including the potentially
concave beginning of the late wood profile to the iadf measure}
}
\value{
a data.frame with "year" in the first column, the measured area
  between a convex approximation and the profile for earlywood ('EW') in the
  second and for latewood ('LW') in the third column.
}
\description{
Get an iadf measure by fitting convex hulls to density profiles
  of the individual intra annual earlywood and latewood sections - this
  function works for individual series.
}
\examples{
#not provided in the development version
}

