% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{init_barplot}
\alias{init_barplot}
\alias{lines_barplot}
\alias{rect_barplot}
\title{custom barplots}
\usage{
init_barplot(x = names(y), y, include0 = TRUE, ylab = "", xlab = "",
  main = "barplot", las = 2)

lines_barplot(x = names(y), y, lwd = 3, col = "black")

rect_barplot(x = names(y), y, width = 0.7, col = "black")
}
\arguments{
\item{x, y}{double, x also character possible, if only one argument is provided
it is taken as y and its names as x. If x can be converted to double it is
used as x, if not it is only used as labels}

\item{include0}{logic, if TRUE y axis includes origin}

\item{ylab}{a title for the y axis}

\item{xlab}{a title for the x axis}

\item{main}{an overall title for the plot}

\item{las}{numeric, see \code{\link[graphics]{par}}}

\item{lwd}{line width as used in \code{\link[graphics]{lines}}}

\item{col}{bar color}

\item{width}{specifies the box width (with = 1 leaves no space between the bars)}
}
\description{
a custom barplots separated in in one init functions and two
  different barplot functions so it can easily be added to other plots or
  used in combination with points, lines and so on.
}
\section{Functions}{
\itemize{
\item \code{init_barplot}: draws an empty plot for rect_barplot and
lines_barplot

\item \code{lines_barplot}: is a plotting function based on \code{\link[graphics]{lines}}

\item \code{rect_barplot}: is a plotting function based on \code{\link[graphics]{rect}}
}}
\examples{
x <- 10 + rnorm(40)
names(x) <- 1901:1940

# rect_barplot:
init_barplot(x)
rect_barplot(x, width = 0.5)

# lines_barplot:
init_barplot(x)
lines_barplot(x, lwd = 3)
}

