% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{intersect_all}
\alias{intersect_all}
\title{intersect_all}
\usage{
intersect_all(a, b, ...)
}
\arguments{
\item{a}{a vector}

\item{b}{a vector}

\item{...}{even more vectors}
}
\description{
a generalization of intersect() to take more than two vectors as
  input
}
\examples{
intersect_all(c(1,2,3,4), c(3,4,5), c(3,4,6,7))
}

