% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgsub}
\alias{mgsub}
\title{generalization of gsub}
\usage{
mgsub(myrepl, mystring)
}
\arguments{
\item{myrepl}{a list containing character vectors of length two each with the
string to be replaced at the first and the replacement at the second index}

\item{mystring}{a character vector where matches are sought, or an object
which can be coerced by as.character to a character vector.}
}
\value{
a character string
}
\description{
a generalization of gsub which allows to do multiple
  replacements at once as described in
 \href{http://stackoverflow.com/questions/15253954/replace-multiple-arguments-with-gsub}{stackoverflow}
}
\examples{
mystring = 'This is good'
myrepl = list(c('o', 'a'), c('i', 'n'))
mgsub(myrepl, mystring)
}

