% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dating.R
\name{new_date_end}
\alias{new_date_end}
\title{new_date_end}
\usage{
new_date_end(rwl, date.end)
}
\arguments{
\item{rwl}{a data.frame/rwl object.}

\item{date.end}{a data.frame with series names in the first, and date end
in the second column.}
}
\value{
a data.frame/rwl object
}
\description{
Batch adjust the date end of multiple series.
}
\examples{
library("dplR")
data("gp.rwl")
de <- dplR::rwl.stats(gp.rwl)[ , c(1,3)] #construct a data frame of current date ends
de[1:4, 2] <- 1600 #changing the date end of the first 4 series to 1600
seg.plot(new_date_end(gp.rwl, de))
}

