% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{new_epsg}
\alias{new_epsg}
\title{new_epsg}
\usage{
new_epsg(spatial_object, epsg_new, epsg_old = NULL)
}
\arguments{
\item{spatial_object}{an object of class "Spatial".}

\item{epsg_new}{an EPSG code provided as character string.}

\item{epsg_old}{optional, an EPSG code provided as character string.}
}
\value{
an object of class "Spatial".
}
\description{
Set new proj4string by supplying a EPSG code.
}

