% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationmatrix.R
\name{panel.cor}
\alias{panel.cor}
\title{panel.cor}
\usage{
panel.cor(w, z, ...)
}
\arguments{
\item{w}{variable}

\item{z}{variable}

\item{...}{additional arguments}
}
\description{
adds a bubble plot panel to a correlation matrix
}
\examples{
pairs(mtcars, upper.panel=panel.cor)
}

