% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bericht.R
\name{plot_bericht}
\alias{plot_bericht}
\title{plot_bericht}
\usage{
plot_bericht(filename, encoding = "latin1", set_lwd = 50, multi = 0.7)
}
\arguments{
\item{filename}{a path to a .xls file used for dating overview in the BOKU
tree ring lab.}

\item{encoding}{encoding of the .xls file}

\item{set_lwd}{width of the segments}

\item{multi}{adjusts the height of the output png image}
}
\description{
Produces a segment plot used in standard dating reports of the
  BOKU tree ring lab based on the overview excel table.
}
\examples{
#use the following line to select the file interactively:
\dontrun{
plot_bericht(file.choose())
}
}
