% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_species_table}
\alias{plot_species_table}
\title{Plot species table}
\usage{
plot_species_table()
}
\value{
nothing, produces a side effect.
}
\description{
Plots the information contained in the species dataset.
}
\examples{
plot_species_table()

}
