% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{po_find}
\alias{po_find}
\title{po_find}
\usage{
po_find(rwl, rc, maxpo = NULL, nyrs = NULL, f = NULL, make.plot = T)
}
\arguments{
\item{rwl}{rwl/data.frame object containing the tree ring series}

\item{rc}{existing "regional curve"}

\item{maxpo}{maximal po until which RSS gets calculated}

\item{nyrs}{setting for the ffcsaps function used to smooth the series:
 a number greater than 1, affecting the rigidity of the
 spline.  When \code{\var{f}} is kept constant, a larger
\code{\var{nyrs}} produces a more rigid spline.  Defaults to
\code{length(\var{y})/2}.}

\item{f}{setting for the ffcsaps function used to smooth the series:
a number between 0 and 1 giving the frequency response at a
wavelength of \code{\var{nyrs}} years.  When \code{\var{nyrs}} is
kept constant, a smaller \code{\var{f}} produces a more rigid
spline: At one extreme, \code{\var{f} = 0} causes the function to
return the least-squares straight line fit to the data.  At the
other extreme, \code{\var{f} = 1} results in the natural spline,
i.e. the function outputs \code{\var{y}}.  The default value is
0.5.}

\item{make.plot}{a \code{logical} indicating if a plot should be drawn}
}
\value{
a \code{data.frane} with the columns "series" and "po", containing
  the series names and po estimations
}
\description{
Estimate pith offset of series by finding the position with
  minimum RSS to existing regional curve
}
\examples{
#no examples added in the current development version - will be
  added in future
}

