% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{po_transform}
\alias{po_transform}
\title{po_transform}
\usage{
po_transform(po, rwl, nyrs = 4)
}
\arguments{
\item{po}{\code{data.frame} of measured distance po (in the same unit as rwl!)
with series names in the first column and po in the second}

\item{rwl}{rwl object containing series measured}

\item{nyrs}{the first \code{1:nyrs} years will be used to calculate the mean growth
rate, default is 4}
}
\value{
a \code{data.frame} just as the input "po" with po as no. of tree rings
}
\description{
estimate no. of missing rings by estimated distance to pith
}
\examples{
no examples available in the development version
}

