% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{qqfunc}
\alias{qqfunc}
\title{visually test normality assumtion}
\usage{
qqfunc(model)
}
\arguments{
\item{model}{a model, output of \code{\link{lm}}}
}
\value{
the number of the plot using real data
}
\description{
try to spot the qqplot of your data within 8 other qqplots,
  drawn randomly from normal distributions as suggested here:
  \url{http://www.nate-miller.org/blog/how-normal-is-normal-a-q-q-plot-approach}
}
\examples{
qqfunc(lm(speed ~ dist, data = cars))
}

