% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono-quality.R
\name{radius_class_plot}
\alias{radius_class_plot}
\title{Plot median growth rate or other parameters per radius class}
\usage{
radius_class_plot(dat, rwdata, cutoff = NULL, classborders = c(0, 50,
  80, 100, 150, 200, Inf), main = "", unit = "[mm]")
}
\arguments{
\item{dat}{data of the investigated tree ring parameter (ring width,
earlywood width, leatewood with, ring density, etc.).}

\item{rwdata}{ring width data to approximate the radius of individual trees.}

\item{cutoff}{cut the data at given year - if NULL the whole range is used.}

\item{classborders}{radii to set the borders of the radius classes.}

\item{main}{plot title.}

\item{unit}{y-axis label.}
}
\description{
Plot median growth rate or other parameters per radius class to
  evaluate trends in tree ring data and data composition.
}
\examples{
#not available in development version.
}
