% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwl-utils.R
\name{radius_rwl}
\alias{radius_rwl}
\title{radius_rwl}
\usage{
radius_rwl(rwl)
}
\arguments{
\item{rwl}{a rwl/data.frame object}
}
\value{
a rwl/data.frame object with cumulated tree ring widths for each
  series (column)
}
\description{
A function to get the cumulated tree ring widths
  (approx. radius of the tree in the given year).
}
\examples{
library("dplR")
data("gp.rwl")
radius_rwl(gp.rwl)
}

