% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_windendro}
\alias{read_windendro}
\title{Read WinDENDRO files}
\usage{
read_windendro(file, encoding = "latin1", trust_ringcount = FALSE)
}
\arguments{
\item{file}{path to the WinDENDRO file}

\item{encoding}{encoding of the WinDENDRO file (should be "latin1")}

\item{trust_ringcount}{if FALSE: values not beeing metadata (= measured values) are checked whether they are of the same length as the ring count specified in the metadata. If commends or marked features are used in WinDENDRO this has to be set to TRUE, then this check is omitted and only the number of values up to the specified ring count read in.}
}
\value{
an object of class `rwl`
}
\description{
Read WinDENDRO files
}
