% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescomp.R
\name{res_comp}
\alias{res_comp}
\title{resilience components}
\usage{
res_comp(rwl, nyrs_pre = 4, nyrs_post = 4)
}
\arguments{
\item{rwl}{a data frame of tree ring measures.}

\item{nyrs_pre}{number of years defining the pre disturbance performance.}

\item{nyrs_post}{number of years defining the post disturbance performance.}
}
\value{
a list of data frames with the members resistance, recovery,
  resilience and rel_resilience. Each data frame consists of the same series
  as supplied with rwl row numbers reduced by nyrs_pre and nyrs_post.
}
\description{
The function calculates resilience components on a data.frame of
 tree-ring series after Lloret et al. (2011), similar to function
 \code{\link[pointRes]{res.comp}}. This function also accepts pre- and post-
 disturbance periods of 1 in contrast to the above mentioned function.
}
\examples{
library('dplR')
data('ca533')
res_comp(ca533)
}

