% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iadf.R
\name{rescale_series_profiles}
\alias{rescale_series_profiles}
\title{rescale density profiles}
\usage{
rescale_series_profiles(data, len = 100)
}
\arguments{
\item{data}{a nested list with the members $profile.raw, $limites and $years
for each individual series (named members on the first hierarchy level)}

\item{len}{the output length of each individual year - usually set to 100.}
}
\value{
a nested list with only the members $profile.raw, $limites and $years
  after the named first hierarchy level.
}
\description{
This function rescales density profiles to the range 0:1 in y
  direction and 1:len in x direction in each individual year.
}
\examples{
#not provided in development version
}

