% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_down}
\alias{round_down}
\title{round_down}
\usage{
round_down(x, to = 1000)
}
\arguments{
\item{x}{numeric or integer}

\item{to}{numeric or integer}
}
\description{
round down a number to the nearest multiple of the argument "to"
}
\examples{
round_down(1412, to = 1000)
# [1] 1000
}

