% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_up}
\alias{round_up}
\title{round_up}
\usage{
round_up(x, to = 1000)
}
\arguments{
\item{x}{numeric or integer}

\item{to}{numeric or integer}
}
\description{
round a number up to the nearest multiple of the argument "to"
}
\examples{
round_up(1412, to = 1000)
# [1] 2000
}

