% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-rwl.R
\name{rwl_plot}
\alias{rwl_plot}
\title{rwl_plot}
\usage{
rwl_plot(rwl, col = "black", ...)
}
\arguments{
\item{rwl}{an object of class dataframe /rwl}

\item{col}{character string indicating the color(s) tu use for the lines}

\item{...}{otional further plotting arguments}
}
\value{
side effect - a plot
}
\description{
useful defaults for \code{\link[graphics]{matplot}} to plot a number of treering series
}
