% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-rwl.R
\name{scatter_rwl}
\alias{scatter_rwl}
\title{scatter_rwl}
\usage{
scatter_rwl(rwl, ...)
}
\arguments{
\item{rwl}{a data frame (potentially containing tree ring series / chronologies).}

\item{...}{additional plotting arguments (e.g. \code{pch = 19}).}
}
\description{
Produces a scatter plot matrix with rainbow colored points as
  well as a panel ob correlation bubbles (size corresponds to squared
  correlation coef, blue positive, red negative)
}
\examples{
library('dplR')
data('gp.rwl')
scatter_rwl(gp.rwl[1:6], main = 'example plot', pch=19)
}

