% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-rwl.R
\name{seg_plot}
\alias{seg_plot}
\title{segment plot}
\usage{
seg_plot(
  rwl,
  col = 1,
  lwd = 4,
  cex = 1,
  xlab = "",
  order = c("first", "last"),
  ...
)
}
\arguments{
\item{rwl}{a data frame resp. object of class `rwl` of ring width series}

\item{col}{color vector with length 1 or the number of series in rwl}

\item{lwd}{line width of the segments}

\item{cex}{size of the axis label and tick labels}

\item{xlab}{character - x axis label}

\item{order}{order of the tree-ring series - according to the year of the first or last ring}

\item{...}{additional parameters for the call of plot()}
}
\value{
side effect - a plot
}
\description{
produces a segment plot of tree-ring data.
}
