% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_range}
\alias{seq_range}
\title{seq_range}
\usage{
seq_range(x, by = 1)
}
\arguments{
\item{x}{a range, numeric vector of lenth 2}

\item{by}{numeric of length 1}
}
\value{
a numeric vector
}
\description{
This function is a user friendly wrapper around seq() and
   calculates a sequence by a given range
}
\examples{
seq_range(c(2, 4), by = 1)
}

