% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{series_length}
\alias{series_length}
\title{series length}
\usage{
series_length(x)
}
\arguments{
\item{x}{a data.frame/rwl object}
}
\value{
a numeric vector
}
\description{
returns the series length of the series within a data.frame/rwl
  object.
}
