% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrending.R
\name{sf_rcs}
\alias{sf_rcs}
\title{signal free rcs}
\usage{
sf_rcs(rwl, po, maxit = Inf)
}
\arguments{
\item{rwl}{a rwl/data.frame object}

\item{po}{a data frame containing series names in the first and po
data as nr. of years in the second column.}

\item{maxit}{maximal number of iterations to take - the iterations stop if
the all values of the SF residual chronology are below 0.002 as suggested
by \href{http://dx.doi.org/10.1016/j.dendro.2013.06.002}{Melvin (2013)} or
reach maxit - default to \code{Inf}}
}
\value{
a list containing:
  \describe{
    \item{rwi}{the indexed tree ring data as data.frame}
    \item{rc}{the final rc curve as numeric vector}
  }
}
\description{
signal free rcs detrending as described by
  \href{http://dx.doi.org/10.1016/j.dendro.2013.06.002}{Melvin (2013)} using
  parts of the code of \code{\link[dplR]{rcs}} as well as the whole function
  internally.
}
\examples{
#not available in development version
}
\references{
Thomas M. Melvin, Keith R. Briffa, CRUST: Software for the
  implementation of Regional Chronology Standardisation: Part 1. Signal-Free
  RCS, Dendrochronologia, Volume 32, Issue 1, 2014, Pages 7-20, ISSN 1125-7865,
  \url{http://dx.doi.org/10.1016/j.dendro.2013.06.002}.

  Thomas M. Melvin, Keith R. Briffa, A “signal-free” approach to
  dendroclimatic standardisation, Dendrochronologia, Volume 26, Issue 2,
  1 October 2008, Pages 71-86, ISSN 1125-7865,
  \url{http://dx.doi.org/10.1016/j.dendro.2007.12.001}.
}

