% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{sort_by_index}
\alias{sort_by_index}
\title{sort_by_index}
\usage{
sort_by_index(x)
}
\arguments{
\item{x}{a numeric vector, representing an individual rwl series,
potentially containing NA values.}
}
\value{
a numeric vector with the same length as x.
}
\description{
internal function sortByIndex from package dplR, shifts
  series to start with index 1, maintaining the same vector length by
  adding NA values to the end. Code copied and not imported to make function
  name consistent to naming convention in package trlboku and to provide
  documentation.
}
\examples{
x <- c(NA,NA,NA,1,2,3,4,5, NA, NA)
sort_by_index(x)
#[1]  1  2  3  4  5 NA NA NA NA NA
}

