% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{species}
\format{
A data.frame with 18 rows and 6 variables:
\describe{
  \item{species}{scientific species name}
  \item{german}{species name in german}
  \item{english}{species name in english}
  \item{idshort}{species id used at stands with more than 99 series}
  \item{idlong}{species id used at stands with less than 99 series}
  \item{color}{color used for plotting}
  }
}
\usage{
species
}
\description{
common species in dendrochronology in central europe, their abbriviations
used in the tree ring lab at BOKU and consistently used colors for plotting.
}
\keyword{datasets}
