% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono-quality.R
\name{stand_depth}
\alias{stand_depth}
\title{stand_depth}
\usage{
stand_depth(crn, rwl, stand = c(1, 3))
}
\arguments{
\item{crn}{chronology object as from dplR::chron().}

\item{rwl}{rwl object.}

\item{stand}{integer vector of length = 2, marking the start and end of
site/stand specification within the series identifier e.g. c(1, 3) for
the stand "Abc" in the series identifier "AbcPA01a".}
}
\value{
a crn object with the column stand.depth added.
}
\description{
This function adds a column containing the "stand_depth" to
  an existing crn object.
}
\examples{
# no example available in the development version
}

