% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono-quality.R
\name{stand_depth_plot}
\alias{stand_depth_plot}
\title{stand_depth_plot}
\usage{
stand_depth_plot(stand_depth_object, chrono = "crn",
  sample.depth = "samp.depth", stand.depth = "stand.depth", main = "",
  col1 = "red", col2 = "orange")
}
\arguments{
\item{stand_depth_object}{an object derived from \code{\link{stand_depth}}.}

\item{chrono}{column name or integer specifying the column index where
the chronology is stored in the stand_depth_object.}

\item{sample.depth}{column name or integer specifying the column index where
the sample depth is stored in the stand_depth_object.}

\item{stand.depth}{column name or integer specifying the column index where
the stand depth is stored in the stand_depth_object.}

\item{main}{optional plot title.}

\item{col1}{optional color for sample depth.}

\item{col2}{optional color for stand depth.}
}
\description{
Plots a chronology with underlaying sample depth and
  "stand depth" information.
}
\examples{
#will be added later
}
