% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidy_series}
\alias{tidy_series}
\title{tidy_series}
\usage{
tidy_series(rwl)
}
\arguments{
\item{rwl}{}
}
\value{
a tibble
}
\description{
transforms an rwl object to tidy data
}
\examples{
library(dplR)
data("ca533")
tidy_series(ca533)
}

