% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidyrwl}
\alias{tidyrwl}
\alias{tidy_crn}
\alias{untidy_crn}
\alias{tidy_rwl}
\alias{untidy_rwl}
\title{tidy and untidy ring width data}
\usage{
tidy_crn(crn)

untidy_crn(tidy_crn)

tidy_rwl(rwl, value_col = "rwl")

untidy_rwl(tidy_rwl, value_col = "rwl")
}
\arguments{
\item{crn}{a chronology as obtained from \code{\link[dplR]{chron}}}

\item{tidy_crn}{a tidy chronology as obtained from \code{\link{tidy_crn}}}

\item{rwl}{ring width data as obtained from \code{\link[dplR]{read.rwl}}}

\item{value_col}{column name of the value column in the tidy tibble of the
input resp output object}

\item{tidy_rwl}{tidy ring width data as obtained from \code{\link{tidy_rwl}}}
}
\value{
data frames or tibbles
}
\description{
little helper functions to convert dataframes from the data
  format used in multiple dendro-related R packages such as \pkg{dplR} to
  tidy data used in the \pkg{tidyverse} and vice versa
}
