% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{transform_projection}
\alias{transform_projection}
\title{transform_projection}
\usage{
transform_projection(spatial_object, epsg_new, epsg_old = NULL)
}
\arguments{
\item{spatial_object}{an object of class "Spatial".}

\item{epsg_new}{an EPSG code provided as character string indicating the
desired projection for the returned object.}

\item{epsg_old}{optional, an EPSG code provided as character string
specifying the projection before transformation if no proj4string is
supplied with the object.}
}
\value{
an object of class "Spatial".
}
\description{
Transforms to different projection by supplying an EPSG code.
  If object has no proj4string defined, old projection can be specified by
  argument epsg_old.

  EPSG of common projections:
  \itemize{
    \item{WG84: 4326}
    \item{Lambert Austria: 31287}
  }
}

