% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{truncate_rwl}
\alias{truncate_rwl}
\title{truncate_rwl}
\usage{
truncate_rwl(x)
}
\arguments{
\item{x}{a data.frame/rwl object}
}
\value{
a data.frame/rwl object
}
\description{
This function removes lines at the beginning and end of a
  data.frame/rwl object only containing NA values (as present after subsetting etc.).
}
\examples{
library("dplR")
data("gp.rwl")
subset <- gp.rwl[ ,1:10] #first value 1720, but rows start with 1570
truncate_rwl(subset) #rows start with 1720
}
