% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{write_fh}
\alias{write_fh}
\alias{write.fh}
\title{Write to a file in Heidelberg format}
\usage{
write_fh(rwl, path, data.format = c("Tree", "HalfChrono", "Chrono"),
  prec = c(0.01, 0.001), meta = NULL, append = FALSE)
}
\arguments{
\item{rwl}{a data.frame/rwl object such as returned e.g. by
\code{\link[dplR]{as.rwl}}.}

\item{path}{a character vector specifying the path to the file where the
data will be written, usually ending with ".fh".}

\item{data.format}{DataFormat as specified by the definition of the
Heidelberg format (either Tree, HalfChrono or Chrono). This is used
for all series within the written object.}

\item{prec}{numeric, indicating the precision of the output file. This must
be equal to either 0.01 or 0.001 (units are in mm). This is used
for all series within the written object.}

\item{meta}{optional additional meta data. If a named vector is supplied,
name value pairs are used as metadata for all series. Supply a list of named
vectors to use different metadata for each individual series.}

\item{append}{logical, indicating whether data should be appended to an
existing file. Defaults to create a new file, overwriting a file in case
it already exists.}
}
\description{
\code{write_fh}, just as its \link[dplR]{dplR} consistent
equivalent \code{write.fh}, writes an rwl object to a file in Heidelberg
(.fh) format. It currently only supports "Single format".
}
