% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yr.range}
\alias{yr.range}
\title{yr.range}
\usage{
yr.range(x, yr.vec = as.numeric(names(x)))
}
\arguments{
\item{x}{a numeric vector}

\item{yr.vec}{a character- or numeric vector containing years with the same
length as x}
}
\value{
a character vector of length 2 containing the first and last year
}
\description{
function yr.range from dplR, extracts the first and last year in
  yr.vec where the corresponding numeric vector shows a value different to NA
}
\examples{
library(dplR)
data(ca533)
yr.range(ca533[1], rownames(ca533))
#[1] "1530" "1983"
}

