% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-rwl.R
\name{expand_apply}
\alias{expand_apply}
\title{expand_apply}
\usage{
expand_apply(x, ...)
}
\arguments{
\item{x}{a numeric vector (NA is allowed and will be omitted)
or a data.frame/rwl object}

\item{FUN}{character, name of a function e.g. 'median'.}
}
\value{
The form of the value depends on the class of its argument. vector
  for default method, data.frame for data.frame method.
}
\description{
Apply function on expanding window.
}
\examples{
#example for numeric method:
x <- c(NA, NA, 1, 3, 2, 3, NA)
expand_apply(x, 'median')
#example for data.frame method:
library('dplR')
data('ca533')
expand_apply(ca533, 'median')
}

